Update .gitignore to exclude database files, Tauri app data, game backups, and Ollama temporary files. Enhance CHANGELOG with new features for translation management, including dedicated translation page and improved UI components. Refactor project loading logic to prevent unnecessary database queries. Clean up deprecated stores and optimize database operations for better performance.

+++ src-tauri/src/commands/scanning.rs
@@ -16,7 +16,7 @@ pub struct ScanState {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ScanRequest {
-    pub project_id: i32,
+    pub project_id: i64,
     pub folder_path: String,
     pub recursive: Option<bool>,
 }
